create wrapper script and update builddags for new config style
